Added fit utility. Added alternate bulk mod calc method.
Murnaghan_fit.f is a fortran program designed to take pairs of Etotal vs
volume for a unit cell and use interpolation to calculate several
physical parameters. Added to 'etot_vs_acell_fitted.ods' is a section
using a 9-point numerical calculation of the 2nd derivative of Etotal
w.r.t. acell. This is then used to find the bulk modulus. It is
currently off by a factor of 2 for some reason.