psp refactoring and beginning of BTO run input file
Moved 3 current psps out of the BTO project and into a dedicated psp
directory in the top level of the repo. That should make it easier to
prevent duplication of psps across projects. Also added a psp library
that can be unzipped to get psps for new atoms as needed.