Begin new approach to optimize BTO struct
Been having problems with SCF cycle convergence that seem to be a bug in the code. Updated Abinit from 7.2.2 to 7.4.1, but it still has the same problem. I posted on the Abinit forum about the issue and saved a copy of the files from the run I used on the forum post. I'm going to try some different approaches now using more/less symmetry and hand selected kpts.