Bug persists. Asking for expert opinion.
Added new psps and tried several different strategems to make BTO code work. It fails with SCF cycle not converging, but also not attempting the requested number of iterations to achieve convergence.

Tried running on the simpler STO structure, but it failed with the same problem. Sent email to some Abinit users for help. Need input/output files to share with them.